% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_append_table}
\alias{db_append_table}
\title{db_append_table}
\usage{
db_append_table(schema, table, data, conn, dryrun = FALSE)
}
\arguments{
\item{schema}{DB schema}

\item{data}{Data values to use for update}

\item{conn}{Existing DBI connection from \code{fieldsets::db_connect}}

\item{dryrun}{Do not write to DB if TRUE}

\item{table_name}{Name of table to append to}
}
\description{
Append a data table to a given data base table.
}
\examples{
db_append_table("public", "my_table", my_data_table, db_connection)
}
