% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_connect}
\alias{db_connect}
\title{db_get_connection}
\usage{
db_connect()
}
\value{
DBI Connection
}
\description{
Return a DBI connection to our Postgres Database
}
\examples{
db_connection <- fieldsets_db_connect
}
