% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_disconnect}
\alias{db_disconnect}
\title{db_disconnect}
\usage{
db_disconnect(conn)
}
\arguments{
\item{conn}{Existing DBI connection from \code{fieldsets::db_connect}}
}
\description{
End a given DBI connection to our Postgres Database
}
\examples{
db_disconnect(db_connection)
}
