% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_fetch_result}
\alias{db_fetch_result}
\title{db_fetch_result}
\usage{
db_fetch_result(query, conn)
}
\arguments{
\item{query}{SQL Query Statement}

\item{conn}{Existing DBI connection from \code{fieldsets::db_connect}}
}
\value{
data frame
}
\description{
Execute a query and return results as a data frame.
}
\examples{
db_fetch_results('SELECT * FROM table;', db_connection)
}
