% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{update_event_status}
\alias{update_event_status}
\title{update_event_status}
\usage{
update_event_status(event_id, new_status, conn)
}
\arguments{
\item{event_id}{The id of the event to update}

\item{new_status}{The new status of the event}

\item{conn}{Existing DBI connection from \code{fieldsets::db_connect}}
}
\description{
Update an event status
}
\examples{
db_connection <- fieldsets_db_connect
}
